requirements.txt
requests
python-dotenv
pandas
pyautogui
screeninfo
